refactor(submissions): A7 — extract DropRepoQueueWidget (P1)
Pure refactor of DropRepoPage.tsx (624 -> 466 lines) — zero behavior
change, visual output identical pre/post.

Moved into new src/components/submissions/DropRepoQueueWidget.tsx:
- Aside "Queue signals" section (Sparkles icon + MetricCard grid)
- Aside "Recent submissions" section (Megaphone icon + submissions list)
- MetricCard local helper (co-located with widget)
- statusLabel switch for SubmissionStatus
- Type definitions: DropRepoQueueSummary, DropRepoPublicSubmission,
  DropRepoSubmissionStatus (re-aliased in DropRepoPage as QueueSummary /
  PublicRepoSubmission / SubmissionStatus to keep existing call-sites
  untouched)

DropRepoPage now imports DropRepoQueueWidget and renders it with the
same props the inline JSX already read (queue, submissions, loading).
No persisted-field or form-behavior change — that work lands in A7-P3.

Verification (clean):
- npm run typecheck
- npm run lint:guards
- npm run test:hooks (42 files / 334 tests)
